# Use zeropool.Pool to workaround SA6002