I built[^1] a tiny library called [zeropool](https://github.com/colega/zeropool) to workaround the SA6002 staticheck issue.

While searching for the references of that SA6002 staticheck issues on Github first results was Prometheus itself, with quite a lot of ignores of it.

This changes the usages of `sync.Pool` to `zeropool.Pool[T]` where a pointer is not available.

Also added a benchmark for HeadAppender Append/Commit when series already exist, which is one of the most usual cases IMO, as I didn't find any.

I think the code is cleaner now, and the benchmarks results are quite promising:

```
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/tsdb
cpu: 11th Gen Intel(R) Core(TM) i7-11700K @ 3.60GHz
                                                                                                         │      old      │                 new                 │
                                                                                                         │    sec/op     │   sec/op     vs base                │
LoadWAL/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=0,mmappedChunkT=0-16         250.7m ±  1%   235.7m ± 0%   -5.97% (p=0.002 n=10)
LoadWAL/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=36,mmappedChunkT=0-16        253.5m ±  3%   237.1m ± 1%   -6.47% (p=0.000 n=10)
LoadWAL/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=72,mmappedChunkT=0-16        256.9m ±  5%   241.2m ± 1%   -6.14% (p=0.000 n=10)
LoadWAL/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=360,mmappedChunkT=0-16       287.6m ±  4%   271.2m ± 1%   -5.73% (p=0.000 n=10)
LoadWAL/batches=10,seriesPerBatch=10000,samplesPerSeries=50,exemplarsPerSeries=0,mmappedChunkT=0-16         357.7m ±  3%   303.9m ± 1%  -15.04% (p=0.000 n=10)
LoadWAL/batches=10,seriesPerBatch=10000,samplesPerSeries=50,exemplarsPerSeries=2,mmappedChunkT=0-16         374.4m ±  4%   324.1m ± 1%  -13.45% (p=0.000 n=10)
LoadWAL/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=0-16         97.35m ± 17%   88.43m ± 1%   -9.17% (p=0.000 n=10)
LoadWAL/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=0-16        124.87m ± 33%   90.69m ± 1%  -27.37% (p=0.000 n=10)
LoadWAL/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=0-16        106.26m ±  7%   93.65m ± 1%  -11.86% (p=0.000 n=10)
LoadWAL/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=0-16        127.4m ± 13%   112.5m ± 1%  -11.63% (p=0.000 n=10)
LoadWAL/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=3800-16      1.151 ±  7%    1.038 ± 1%   -9.77% (p=0.000 n=10)
LoadWAL/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=3800-16      1.146 ± 17%    1.077 ± 2%   -6.03% (p=0.005 n=10)
LoadWAL/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=3800-16      1.140 ± 42%    1.113 ± 3%        ~ (p=0.247 n=10)
LoadWAL/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=3800-16     1.315 ± 20%    1.280 ± 1%   -2.69% (p=0.002 n=10)
geomean                                                                                                     332.8m         300.3m        -9.78%

                                                                                                         │     old      │                 new                 │
                                                                                                         │     B/op     │     B/op      vs base               │
LoadWAL/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=0,mmappedChunkT=0-16        38.95Mi ± 5%   37.28Mi ± 0%  -4.27% (p=0.000 n=10)
LoadWAL/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=36,mmappedChunkT=0-16       42.65Mi ± 4%   39.12Mi ± 0%  -8.27% (p=0.000 n=10)
LoadWAL/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=72,mmappedChunkT=0-16       44.33Mi ± 0%   40.79Mi ± 0%  -7.98% (p=0.000 n=10)
LoadWAL/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=360,mmappedChunkT=0-16      57.71Mi ± 0%   55.98Mi ± 0%  -2.99% (p=0.000 n=10)
LoadWAL/batches=10,seriesPerBatch=10000,samplesPerSeries=50,exemplarsPerSeries=0,mmappedChunkT=0-16        230.5Mi ± 2%   227.6Mi ± 2%       ~ (p=0.165 n=10)
LoadWAL/batches=10,seriesPerBatch=10000,samplesPerSeries=50,exemplarsPerSeries=2,mmappedChunkT=0-16        259.2Mi ± 2%   249.4Mi ± 5%  -3.79% (p=0.001 n=10)
LoadWAL/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=0-16        44.44Mi ± 3%   43.36Mi ± 2%  -2.43% (p=0.000 n=10)
LoadWAL/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=0-16        47.79Mi ± 4%   44.94Mi ± 1%  -5.96% (p=0.000 n=10)
LoadWAL/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=0-16        48.77Mi ± 2%   46.28Mi ± 2%  -5.11% (p=0.000 n=10)
LoadWAL/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=0-16       58.40Mi ± 1%   56.54Mi ± 1%  -3.18% (p=0.000 n=10)
LoadWAL/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=3800-16    293.5Mi ± 3%   278.9Mi ± 4%  -4.99% (p=0.001 n=10)
LoadWAL/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=3800-16    301.1Mi ± 2%   290.3Mi ± 3%  -3.59% (p=0.002 n=10)
LoadWAL/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=3800-16    307.2Mi ± 8%   305.4Mi ± 3%       ~ (p=0.123 n=10)
LoadWAL/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=3800-16   398.0Mi ± 3%   393.6Mi ± 1%  -1.11% (p=0.023 n=10)
geomean                                                                                                    103.7Mi        99.54Mi       -3.99%

                                                                                                         │     old     │                 new                 │
                                                                                                         │  allocs/op  │  allocs/op   vs base                │
LoadWAL/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=0,mmappedChunkT=0-16        514.5k ± 3%   442.2k ± 0%  -14.05% (p=0.000 n=10)
LoadWAL/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=36,mmappedChunkT=0-16       636.9k ± 2%   550.6k ± 0%  -13.54% (p=0.000 n=10)
LoadWAL/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=72,mmappedChunkT=0-16       745.7k ± 0%   659.0k ± 0%  -11.62% (p=0.000 n=10)
LoadWAL/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=360,mmappedChunkT=0-16      1.616M ± 0%   1.540M ± 0%   -4.68% (p=0.000 n=10)
LoadWAL/batches=10,seriesPerBatch=10000,samplesPerSeries=50,exemplarsPerSeries=0,mmappedChunkT=0-16        2.194M ± 0%   2.193M ± 0%        ~ (p=0.105 n=10)
LoadWAL/batches=10,seriesPerBatch=10000,samplesPerSeries=50,exemplarsPerSeries=2,mmappedChunkT=0-16        2.795M ± 0%   2.792M ± 0%   -0.12% (p=0.001 n=10)
LoadWAL/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=0-16        435.7k ± 1%   427.7k ± 0%   -1.84% (p=0.000 n=10)
LoadWAL/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=0-16        502.7k ± 2%   487.2k ± 0%   -3.07% (p=0.000 n=10)
LoadWAL/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=0-16        589.1k ± 0%   577.9k ± 0%   -1.90% (p=0.000 n=10)
LoadWAL/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=0-16       1.163M ± 0%   1.154M ± 0%   -0.77% (p=0.000 n=10)
LoadWAL/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=3800-16    3.473M ± 1%   3.394M ± 1%   -2.26% (p=0.000 n=10)
LoadWAL/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=3800-16    4.067M ± 0%   3.995M ± 0%   -1.77% (p=0.000 n=10)
LoadWAL/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=3800-16    4.948M ± 1%   4.897M ± 0%   -1.02% (p=0.000 n=10)
LoadWAL/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=3800-16   10.66M ± 0%   10.60M ± 0%   -0.54% (p=0.000 n=10)
geomean                                                                                                    1.492M        1.429M        -4.22%
```

```
                                             │     old     │                new                 │
                                             │   sec/op    │   sec/op     vs base               │
HeadAppender_Append_Commit_ExistingSeries-16   343.0µ ± 2%   313.6µ ± 1%  -8.58% (p=0.000 n=10)

                                             │     old      │                 new                 │
                                             │     B/op     │     B/op      vs base               │
HeadAppender_Append_Commit_ExistingSeries-16   53.11Ki ± 0%   52.72Ki ± 0%  -0.74% (p=0.000 n=10)

                                             │    old     │                new                │
                                             │ allocs/op  │ allocs/op   vs base               │
HeadAppender_Append_Commit_ExistingSeries-16   624.5 ± 0%   617.0 ± 0%  -1.20% (p=0.000 n=10)
```

[^1]: Shameless self plug.