@colega 2023-03-29
I have pushed two more commits here: 
- [one to make sure that the pool_test.go isn't flaky](https://github.com/prometheus/prometheus/pull/12189/commits/c33bba16c867eba11b1fe3b05463c94a0bb1f4c2) (because since it runs as part of a bigger codebase here, GC could kick in and pools would be flushed, causing an allocation).
- Another one changing the last reference to SA6002 I found, in `engine.go`, it took me a while to run the benchmarks for this, which I include collsapsed below.

@bwplotka I'd appreciate one more check of those.

<details>
<summary>BenchmarkRangeQuery results</summary>

```
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/promql
cpu: Intel(R) Xeon(R) CPU @ 3.10GHz
BenchmarkRangeQuery/expr=a_one,steps=1-8         	   91327	     12281 ns/op	    5789 B/op	     116 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=1-8         	   91711	     12262 ns/op	    5788 B/op	     116 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=1-8         	   92962	     12240 ns/op	    5788 B/op	     116 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=1-8         	   95372	     12192 ns/op	    5787 B/op	     116 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=1-8         	   96512	     12180 ns/op	    5788 B/op	     116 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=1-8         	   96238	     12268 ns/op	    5788 B/op	     116 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=1-8         	   95694	     12232 ns/op	    5788 B/op	     116 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=1-8         	   95006	     12317 ns/op	    5788 B/op	     116 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=1-8         	   93148	     12328 ns/op	    5788 B/op	     116 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=1-8         	   94510	     12253 ns/op	    5788 B/op	     116 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=10-8        	   94050	     12524 ns/op	    5788 B/op	     116 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=10-8        	   92900	     12503 ns/op	    5788 B/op	     116 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=10-8        	   92430	     12464 ns/op	    5788 B/op	     116 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=10-8        	   93544	     12584 ns/op	    5788 B/op	     116 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=10-8        	   92546	     12530 ns/op	    5789 B/op	     116 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=10-8        	   91596	     12519 ns/op	    5788 B/op	     116 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=10-8        	   92529	     12603 ns/op	    5788 B/op	     116 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=10-8        	   92761	     12499 ns/op	    5788 B/op	     116 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=10-8        	   93618	     12553 ns/op	    5788 B/op	     116 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=10-8        	   92350	     12464 ns/op	    5788 B/op	     116 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=100-8       	   66008	     17630 ns/op	    6029 B/op	     119 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=100-8       	   66109	     17620 ns/op	    6029 B/op	     119 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=100-8       	   65830	     17669 ns/op	    6029 B/op	     119 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=100-8       	   65946	     17595 ns/op	    6029 B/op	     119 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=100-8       	   65798	     17652 ns/op	    6029 B/op	     119 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=100-8       	   66628	     17668 ns/op	    6029 B/op	     119 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=100-8       	   66292	     17614 ns/op	    6029 B/op	     119 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=100-8       	   66420	     17654 ns/op	    6030 B/op	     119 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=100-8       	   66529	     17650 ns/op	    6030 B/op	     119 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=100-8       	   66752	     17713 ns/op	    6029 B/op	     119 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=1000-8      	   21549	     54908 ns/op	    8961 B/op	     139 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=1000-8      	   21607	     55036 ns/op	    8965 B/op	     139 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=1000-8      	   21691	     54732 ns/op	    8964 B/op	     139 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=1000-8      	   21699	     54446 ns/op	    8968 B/op	     139 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=1000-8      	   21736	     54811 ns/op	    8967 B/op	     139 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=1000-8      	   21613	     55001 ns/op	    8967 B/op	     139 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=1000-8      	   21699	     54814 ns/op	    8964 B/op	     139 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=1000-8      	   21638	     54743 ns/op	    8967 B/op	     139 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=1000-8      	   21676	     54801 ns/op	    8966 B/op	     139 allocs/op
BenchmarkRangeQuery/expr=a_one,steps=1000-8      	   21781	     54930 ns/op	    8970 B/op	     139 allocs/op
BenchmarkRangeQuery/expr=a_ten,steps=1-8         	   30646	     38808 ns/op	   11304 B/op	     201 allocs/op
BenchmarkRangeQuery/expr=a_ten,steps=1-8         	   30566	     38673 ns/op	   11303 B/op	     201 allocs/op
BenchmarkRangeQuery/expr=a_ten,steps=1-8         	^Coleg@oleg-benchmarks:~/prometheus$ go test -timeout=240m -run=X -bench="BenchmarkRangeQuery" -count=10 -benchmem ./promql | tee BenchmarkRangeQuery.old^C
oleg@oleg-benchmarks:~/prometheus$ tmux
[exited]
oleg@oleg-benchmarks:~/prometheus$ tmux
[exited]
oleg@oleg-benchmarks:~/prometheus$ go test -timeout=240m -run=X -bench="BenchmarkRangeQuery" -count=10 -benchmem ./promql | tee BenchmarkRangeQuery.old^C
oleg@oleg-benchmarks:~/prometheus$ tmux
[exited]
oleg@oleg-benchmarks:~/prometheus$ 
oleg@oleg-benchmarks:~/prometheus$ ls
BenchmarkRangeQuery.new  CODE_OF_CONDUCT.md  LICENSE         Makefile.common  RELEASE.md   cmd                consoles   documentation  model     plugins      promql  scripts   tracing  web
BenchmarkRangeQuery.old  CONTRIBUTING.md     MAINTAINERS.md  NOTICE           SECURITY.md  config             discovery  go.mod         notifier  plugins.yml  rules   storage   tsdb
CHANGELOG.md             Dockerfile          Makefile        README.md        VERSION      console_libraries  docs       go.sum         patch     prompb       scrape  template  util
oleg@oleg-benchmarks:~/prometheus$ cp BenchmarkRangeQuery.old ../old
oleg@oleg-benchmarks:~/prometheus$ cp BenchmarkRangeQuery.new ../new
oleg@oleg-benchmarks:~/prometheus$ cd ..
oleg@oleg-benchmarks:~$ go/bin/benchstat old new 
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/promql
cpu: Intel(R) Xeon(R) CPU @ 3.10GHz
                                                                                                         │     old      │                new                 │
                                                                                                         │    sec/op    │   sec/op     vs base               │
RangeQuery/expr=a_one,steps=1-8                                                                             12.27µ ± 0%   12.37µ ± 1%  +0.82% (p=0.005 n=10)
RangeQuery/expr=a_one,steps=10-8                                                                            12.56µ ± 0%   12.61µ ± 1%  +0.41% (p=0.023 n=10)
RangeQuery/expr=a_one,steps=100-8                                                                           18.07µ ± 0%   17.66µ ± 1%  -2.24% (p=0.000 n=10)
RangeQuery/expr=a_one,steps=1000-8                                                                          58.08µ ± 0%   54.36µ ± 0%  -6.42% (p=0.000 n=10)
RangeQuery/expr=a_ten,steps=1-8                                                                             39.01µ ± 0%   38.97µ ± 0%       ~ (p=0.631 n=10)
RangeQuery/expr=a_ten,steps=10-8                                                                            41.73µ ± 0%   41.39µ ± 0%  -0.81% (p=0.000 n=10)
RangeQuery/expr=a_ten,steps=100-8                                                                           91.46µ ± 1%   90.58µ ± 0%  -0.97% (p=0.000 n=10)
RangeQuery/expr=a_ten,steps=1000-8                                                                          483.9µ ± 0%   451.3µ ± 0%  -6.73% (p=0.000 n=10)
RangeQuery/expr=a_hundred,steps=1-8                                                                         378.6µ ± 1%   370.8µ ± 1%  -2.05% (p=0.000 n=10)
RangeQuery/expr=a_hundred,steps=10-8                                                                        406.2µ ± 0%   397.5µ ± 0%  -2.13% (p=0.000 n=10)
RangeQuery/expr=a_hundred,steps=100-8                                                                       939.6µ ± 0%   914.3µ ± 0%  -2.69% (p=0.000 n=10)
RangeQuery/expr=a_hundred,steps=1000-8                                                                      5.004m ± 0%   4.659m ± 0%  -6.89% (p=0.000 n=10)
RangeQuery/expr=rate(a_one[1m]),steps=1-8                                                                   17.60µ ± 0%   17.37µ ± 1%  -1.30% (p=0.000 n=10)
RangeQuery/expr=rate(a_one[1m]),steps=10-8                                                                  18.65µ ± 1%   18.50µ ± 1%  -0.81% (p=0.000 n=10)
RangeQuery/expr=rate(a_one[1m]),steps=100-8                                                                 32.60µ ± 0%   31.90µ ± 0%  -2.14% (p=0.000 n=10)
RangeQuery/expr=rate(a_one[1m]),steps=1000-8                                                                151.4µ ± 0%   148.3µ ± 0%  -2.07% (p=0.000 n=10)
RangeQuery/expr=rate(a_ten[1m]),steps=1-8                                                                   47.82µ ± 0%   47.64µ ± 1%  -0.38% (p=0.011 n=10)
RangeQuery/expr=rate(a_ten[1m]),steps=10-8                                                                  58.49µ ± 0%   58.42µ ± 0%       ~ (p=0.353 n=10)
RangeQuery/expr=rate(a_ten[1m]),steps=100-8                                                                 191.4µ ± 0%   192.0µ ± 0%  +0.32% (p=0.000 n=10)
RangeQuery/expr=rate(a_ten[1m]),steps=1000-8                                                                1.372m ± 0%   1.348m ± 0%  -1.80% (p=0.000 n=10)
RangeQuery/expr=rate(a_hundred[1m]),steps=1-8                                                               445.3µ ± 1%   442.3µ ± 0%  -0.66% (p=0.000 n=10)
RangeQuery/expr=rate(a_hundred[1m]),steps=10-8                                                              556.9µ ± 0%   552.3µ ± 0%  -0.83% (p=0.000 n=10)
RangeQuery/expr=rate(a_hundred[1m]),steps=100-8                                                             1.920m ± 0%   1.914m ± 0%  -0.33% (p=0.001 n=10)
RangeQuery/expr=rate(a_hundred[1m]),steps=1000-8                                                            13.79m ± 0%   13.50m ± 0%  -2.10% (p=0.000 n=10)
RangeQuery/expr=rate(a_one[1m]),steps=10000-8                                                               1.465m ± 0%   1.456m ± 0%  -0.61% (p=0.000 n=10)
RangeQuery/expr=rate(a_ten[1m]),steps=10000-8                                                               14.38m ± 0%   14.34m ± 0%  -0.23% (p=0.000 n=10)
RangeQuery/expr=rate(a_hundred[1m]),steps=10000-8                                                           146.1m ± 0%   146.3m ± 0%       ~ (p=0.089 n=10)
RangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=1-8                                                 1.131m ± 1%   1.138m ± 1%  +0.63% (p=0.023 n=10)
RangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=10-8                                                1.653m ± 1%   1.671m ± 1%  +1.11% (p=0.002 n=10)
RangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=100-8                                               6.963m ± 0%   6.950m ± 0%       ~ (p=0.089 n=10)
RangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=1000-8                                              59.53m ± 0%   59.64m ± 0%  +0.18% (p=0.002 n=10)
RangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=1-8                                                 6.271m ± 0%   6.315m ± 3%  +0.70% (p=0.000 n=10)
RangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=10-8                                                11.64m ± 0%   11.63m ± 0%       ~ (p=0.393 n=10)
RangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=100-8                                               64.17m ± 0%   64.21m ± 0%       ~ (p=0.393 n=10)
RangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=1000-8                                              585.4m ± 0%   585.3m ± 0%       ~ (p=0.684 n=10)
RangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=1-8                                             58.08m ± 0%   58.36m ± 0%  +0.49% (p=0.000 n=10)
RangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=10-8                                            110.8m ± 0%   111.1m ± 0%  +0.32% (p=0.002 n=10)
RangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=100-8                                           632.3m ± 0%   632.7m ± 0%  +0.08% (p=0.002 n=10)
RangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=1000-8                                           5.848 ± 0%    5.849 ± 0%       ~ (p=0.353 n=10)
RangeQuery/expr=changes(a_one[1d]),steps=1-8                                                                793.5µ ± 1%   800.5µ ± 1%  +0.88% (p=0.009 n=10)
RangeQuery/expr=changes(a_one[1d]),steps=10-8                                                               932.1µ ± 1%   924.3µ ± 0%  -0.84% (p=0.001 n=10)
RangeQuery/expr=changes(a_one[1d]),steps=100-8                                                              2.207m ± 0%   2.071m ± 0%  -6.18% (p=0.000 n=10)
RangeQuery/expr=changes(a_one[1d]),steps=1000-8                                                             14.69m ± 0%   13.31m ± 0%  -9.38% (p=0.000 n=10)
RangeQuery/expr=changes(a_ten[1d]),steps=1-8                                                                5.109m ± 0%   5.094m ± 0%       ~ (p=0.105 n=10)
RangeQuery/expr=changes(a_ten[1d]),steps=10-8                                                               6.412m ± 0%   6.250m ± 0%  -2.54% (p=0.000 n=10)
RangeQuery/expr=changes(a_ten[1d]),steps=100-8                                                              18.98m ± 0%   17.56m ± 0%  -7.51% (p=0.000 n=10)
RangeQuery/expr=changes(a_ten[1d]),steps=1000-8                                                             143.2m ± 0%   129.4m ± 0%  -9.65% (p=0.000 n=10)
RangeQuery/expr=changes(a_hundred[1d]),steps=1-8                                                            48.83m ± 0%   48.71m ± 0%       ~ (p=0.063 n=10)
RangeQuery/expr=changes(a_hundred[1d]),steps=10-8                                                           61.91m ± 0%   60.11m ± 0%  -2.92% (p=0.000 n=10)
RangeQuery/expr=changes(a_hundred[1d]),steps=100-8                                                          187.3m ± 0%   172.9m ± 0%  -7.68% (p=0.000 n=10)
RangeQuery/expr=changes(a_hundred[1d]),steps=1000-8                                                          1.429 ± 0%    1.286 ± 0%  -9.99% (p=0.000 n=10)
RangeQuery/expr=rate(a_one[1d]),steps=1-8                                                                   784.1µ ± 1%   794.4µ ± 1%  +1.32% (p=0.004 n=10)
RangeQuery/expr=rate(a_one[1d]),steps=10-8                                                                  887.9µ ± 1%   896.6µ ± 1%  +0.98% (p=0.000 n=10)
RangeQuery/expr=rate(a_one[1d]),steps=100-8                                                                 1.813m ± 0%   1.823m ± 0%  +0.56% (p=0.000 n=10)
RangeQuery/expr=rate(a_one[1d]),steps=1000-8                                                                10.85m ± 0%   10.87m ± 0%  +0.17% (p=0.001 n=10)
RangeQuery/expr=rate(a_ten[1d]),steps=1-8                                                                   5.014m ± 1%   5.043m ± 0%  +0.58% (p=0.020 n=10)
RangeQuery/expr=rate(a_ten[1d]),steps=10-8                                                                  5.949m ± 0%   5.950m ± 1%       ~ (p=0.853 n=10)
RangeQuery/expr=rate(a_ten[1d]),steps=100-8                                                                 15.04m ± 0%   15.04m ± 0%       ~ (p=0.631 n=10)
RangeQuery/expr=rate(a_ten[1d]),steps=1000-8                                                                104.9m ± 0%   104.8m ± 0%       ~ (p=0.579 n=10)
RangeQuery/expr=rate(a_hundred[1d]),steps=1-8                                                               47.89m ± 0%   48.14m ± 0%  +0.52% (p=0.000 n=10)
RangeQuery/expr=rate(a_hundred[1d]),steps=10-8                                                              57.21m ± 0%   57.36m ± 0%  +0.27% (p=0.019 n=10)
RangeQuery/expr=rate(a_hundred[1d]),steps=100-8                                                             147.7m ± 0%   148.1m ± 0%  +0.27% (p=0.000 n=10)
RangeQuery/expr=rate(a_hundred[1d]),steps=1000-8                                                             1.042 ± 0%    1.042 ± 0%       ~ (p=0.739 n=10)
RangeQuery/expr=absent_over_time(a_one[1d]),steps=1-8                                                       773.5µ ± 2%   784.0µ ± 1%  +1.37% (p=0.000 n=10)
RangeQuery/expr=absent_over_time(a_one[1d]),steps=10-8                                                      816.5µ ± 1%   823.6µ ± 0%  +0.86% (p=0.004 n=10)
RangeQuery/expr=absent_over_time(a_one[1d]),steps=100-8                                                     1.117m ± 0%   1.130m ± 0%  +1.25% (p=0.000 n=10)
RangeQuery/expr=absent_over_time(a_one[1d]),steps=1000-8                                                    3.884m ± 0%   3.890m ± 0%  +0.16% (p=0.043 n=10)
RangeQuery/expr=absent_over_time(a_ten[1d]),steps=1-8                                                       4.889m ± 0%   4.909m ± 0%  +0.42% (p=0.000 n=10)
RangeQuery/expr=absent_over_time(a_ten[1d]),steps=10-8                                                      5.201m ± 0%   5.228m ± 0%  +0.52% (p=0.000 n=10)
RangeQuery/expr=absent_over_time(a_ten[1d]),steps=100-8                                                     8.033m ± 0%   8.048m ± 0%  +0.19% (p=0.005 n=10)
RangeQuery/expr=absent_over_time(a_ten[1d]),steps=1000-8                                                    35.59m ± 0%   35.70m ± 0%  +0.33% (p=0.019 n=10)
RangeQuery/expr=absent_over_time(a_hundred[1d]),steps=1-8                                                   46.48m ± 0%   46.76m ± 0%  +0.59% (p=0.003 n=10)
RangeQuery/expr=absent_over_time(a_hundred[1d]),steps=10-8                                                  49.34m ± 0%   49.62m ± 1%  +0.56% (p=0.002 n=10)
RangeQuery/expr=absent_over_time(a_hundred[1d]),steps=100-8                                                 77.79m ± 0%   78.12m ± 0%  +0.43% (p=0.000 n=10)
RangeQuery/expr=absent_over_time(a_hundred[1d]),steps=1000-8                                                352.5m ± 0%   351.7m ± 0%  -0.22% (p=0.000 n=10)
RangeQuery/expr=-a_one,steps=1-8                                                                            14.21µ ± 1%   14.13µ ± 0%  -0.55% (p=0.004 n=10)
RangeQuery/expr=-a_one,steps=10-8                                                                           14.50µ ± 1%   14.35µ ± 0%  -1.10% (p=0.000 n=10)
RangeQuery/expr=-a_one,steps=100-8                                                                          20.15µ ± 0%   19.41µ ± 0%  -3.66% (p=0.000 n=10)
RangeQuery/expr=-a_one,steps=1000-8                                                                         61.08µ ± 0%   56.77µ ± 0%  -7.05% (p=0.000 n=10)
RangeQuery/expr=-a_ten,steps=1-8                                                                            44.13µ ± 1%   43.83µ ± 0%  -0.70% (p=0.000 n=10)
RangeQuery/expr=-a_ten,steps=10-8                                                                           46.90µ ± 1%   46.45µ ± 0%  -0.95% (p=0.000 n=10)
…